+++ src/util/constant.js
@@ -1,0 +1,12 @@
+export const $ = {
+  LABEL_LIST: 'ul.label-list',
+  LABEL_SHOW_BTN: '.new-label-button',
+  LABEL_HIDE_BTN: '#label-cancel-button',
+  LABEL_FORM: '#new-label-form',
+  INPUT_LABEL_NAME: '#label-name-input',
+  INPUT_LABEL_DESCRIPTION: '#label-description-input',
+  INPUT_LABEL_COLOR: '#label-color-value',
+  LABEL_CREATE_BTN: '#label-create-button',
+  LABEL_COLOR_BTN: '#new-label-color',
+  LABEL_PREVIEW: '#label-preview',
+}
